Back/requirements_simple.txt
Flask
Flask-CORS
numpy
pandas
Pillow
scikit-learn
joblib
